add mailcatcher, remove mailslurper
