--- ci/config/notification-service/testplatform.config.json
@@ -1,17 +1,0 @@
-{
-	"mail": {
-	  "options": {
-		"host": "mailslurper",
-		"port": 2500,
-		"secure": false,
-		"auth": {
-		  "user": "",
-		  "pass": ""
-		}
-	  },
-	  "defaults": {
-		"from": "Sample Service <snodejs.smtp@gmail.com>",
-		"delay": 0
-	  }
-	}
-}
