--- scripts/ci/mailslurper-config.json
@@ -1,18 +1,0 @@
-{
-	"wwwAddress": "0.0.0.0",
-	"wwwPort": 8080,
-	"serviceAddress": "0.0.0.0",
-	"servicePort": 8085,
-	"smtpAddress": "0.0.0.0",
-	"smtpPort": 2500,
-	"dbEngine": "SQLite",
-	"dbHost": "",
-	"dbPort": 0,
-	"dbDatabase": "./mailslurper.db",
-	"dbUserName": "",
-	"dbPassword": "",
-	"maxWorkers": 1000,
-	"autoStartBrowser": false,
-	"keyFile": "",
-	"certFile": ""
-}
